Refactor the Redis/Cockroach interceptors
