+++ src/main/java/cloud/filibuster/junit/configuration/examples/db/byzantine/redis/RedisSingleGetBasicStringByzantineFaultAnalysisConfigurationFile.java
@@ -1,21 +1,21 @@
-package cloud.filibuster.junit.configuration.examples.redis;
+package cloud.filibuster.junit.configuration.examples.db.byzantine.redis;
 
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfiguration;
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfigurationFile;
 import cloud.filibuster.junit.configuration.FilibusterCustomAnalysisConfigurationFile;
+import cloud.filibuster.junit.configuration.examples.db.byzantine.types.ByzantineStringFaultType;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import static cloud.filibuster.instrumentation.Constants.REDIS_MODULE_NAME;
 
-public class RedisSingleFaultCommandInterruptedExceptionAnalysisConfigurationFile implements FilibusterAnalysisConfigurationFile {
+public class RedisSingleGetBasicStringByzantineFaultAnalysisConfigurationFile implements FilibusterAnalysisConfigurationFile {
     private static final FilibusterCustomAnalysisConfigurationFile filibusterCustomAnalysisConfigurationFile;
 
-    private static Map<String, String> createErrorMap() {
-        Map<String, String> myMap = new HashMap<>();
-        myMap.put("cause", "Command interrupted");
-        myMap.put("code", "");
+    private static <T> Map<String, T> createBzyantineFaultMap(T value) {
+        Map<String, T> myMap = new HashMap<>();
+        myMap.put("value", value);
         return myMap;
     }
 
@@ -23,10 +23,14 @@ public class RedisSingleFaultCommandInterruptedExceptionAnalysisConfigurationFil
         FilibusterCustomAnalysisConfigurationFile.Builder filibusterCustomAnalysisConfigurationFileBuilder = new FilibusterCustomAnalysisConfigurationFile.Builder();
 
         FilibusterAnalysisConfiguration.Builder filibusterAnalysisConfigurationBuilderRedisExceptions = new FilibusterAnalysisConfiguration.Builder()
-                .name("java.lettuce.exceptions.RedisCommandInterruptedException")
-                .pattern(REDIS_MODULE_NAME + "/(await)\\b");
+                .name("java.lettuce.byzantine.basic_string")
+                .pattern(REDIS_MODULE_NAME + "/(get)\\b");
 
-        filibusterAnalysisConfigurationBuilderRedisExceptions.exception("io.lettuce.core.RedisCommandInterruptedException", createErrorMap());
+
+        String[] possibleValues = {null, ""};
+        for (String value : possibleValues) {
+            filibusterAnalysisConfigurationBuilderRedisExceptions.byzantine(new ByzantineStringFaultType(), createBzyantineFaultMap(value));
+        }
 
         filibusterCustomAnalysisConfigurationFileBuilder.analysisConfiguration(filibusterAnalysisConfigurationBuilderRedisExceptions.build());
