+++ src/main/java/cloud/filibuster/junit/configuration/examples/db/byzantine/redis/RedisSingleGetByteArrByzantineFaultAnalysisConfigurationFile.java
@@ -1,9 +1,9 @@
-package cloud.filibuster.junit.configuration.examples.redis.byzantine;
+package cloud.filibuster.junit.configuration.examples.db.byzantine.redis;
 
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfiguration;
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfigurationFile;
 import cloud.filibuster.junit.configuration.FilibusterCustomAnalysisConfigurationFile;
-import cloud.filibuster.junit.configuration.examples.redis.byzantine.types.ByzantineByteArrayFaultType;
+import cloud.filibuster.junit.configuration.examples.db.byzantine.types.ByzantineByteArrayFaultType;
 
 import java.nio.charset.Charset;
 import java.util.HashMap;
