+++ src/main/java/cloud/filibuster/junit/configuration/examples/db/byzantine/redis/RedisSingleGetStringByzantineFaultAnalysisConfigurationFile.java
@@ -1,9 +1,9 @@
-package cloud.filibuster.junit.configuration.examples.redis.byzantine;
+package cloud.filibuster.junit.configuration.examples.db.byzantine.redis;
 
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfiguration;
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfigurationFile;
 import cloud.filibuster.junit.configuration.FilibusterCustomAnalysisConfigurationFile;
-import cloud.filibuster.junit.configuration.examples.redis.byzantine.types.ByzantineStringFaultType;
+import cloud.filibuster.junit.configuration.examples.db.byzantine.types.ByzantineStringFaultType;
 
 import java.util.HashMap;
 import java.util.Map;
