+++ src/main/java/cloud/filibuster/junit/configuration/examples/db/byzantine/types/ByzantineFault.java
@@ -1,0 +1,5 @@
+package cloud.filibuster.junit.configuration.examples.db.byzantine.types;
+
+enum ByzantineFault {
+    STRING, BYTE_ARRAY
+}
