+++ src/main/java/cloud/filibuster/junit/configuration/examples/db/byzantine/types/ByzantineFaultType.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.junit.configuration.examples.redis.byzantine.types;
+package cloud.filibuster.junit.configuration.examples.db.byzantine.types;
 
 import cloud.filibuster.exceptions.filibuster.FilibusterRuntimeException;
