--- src/main/java/cloud/filibuster/junit/configuration/examples/redis/byzantine/types/ByzantineFault.java
@@ -1,5 +1,0 @@
-package cloud.filibuster.junit.configuration.examples.redis.byzantine.types;
-
-enum ByzantineFault {
-    STRING, BYTE_ARRAY
-}
