+++ src/test/java/cloud/filibuster/functional/java/redis/JUnitRedisFilibusterByzantineByteArrTest.java
@@ -4,7 +4,7 @@ import cloud.filibuster.functional.java.JUnitAnnotationBaseTest;
 import cloud.filibuster.instrumentation.libraries.lettuce.RedisInterceptorFactory;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.RedisClientService;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.configuration.examples.redis.byzantine.RedisSingleGetByteArrByzantineFaultAnalysisConfigurationFile;
+import cloud.filibuster.junit.configuration.examples.db.byzantine.redis.RedisSingleGetByteArrByzantineFaultAnalysisConfigurationFile;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.sync.RedisCommands;
 import io.lettuce.core.codec.ByteArrayCodec;
